requirements.txt
pyparsing==2.1.0
pytest==3.3.2
pytest-cov==2.5.1
pytest-mock==1.6.3
